+++ core/src/main/java/nl/nn/adapterframework/senders/IbisJavaSender.java
@@ -1,5 +1,5 @@
 /*
-   Copyright 2013 Nationale-Nederlanden
+   Copyright 2013, 2016 Nationale-Nederlanden
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
 */
 package nl.nn.adapterframework.senders;
 
+import java.io.ByteArrayInputStream;
 import java.util.HashMap;
 
 import nl.nn.adapterframework.configuration.ConfigurationException;
@@ -25,6 +26,7 @@ import nl.nn.adapterframework.core.SenderWithParametersBase;
 import nl.nn.adapterframework.core.TimeOutException;
 import nl.nn.adapterframework.dispatcher.DispatcherManager;
 import nl.nn.adapterframework.dispatcher.DispatcherManagerFactory;
+import nl.nn.adapterframework.http.HttpSender;
 import nl.nn.adapterframework.parameters.ParameterResolutionContext;
 import nl.nn.adapterframework.util.Misc;
 
@@ -44,6 +46,9 @@ import org.apache.commons.lang.StringUtils;
  * <tr><td>{@link #setServiceName(String) serviceName}</td><td>serviceName of the 
  * {@link nl.nn.adapterframework.receivers.JavaListener JavaListener} that should be called</td><td>&nbsp;</td></tr>
  * <tr><td>{@link #setReturnedSessionKeys(String) returnedSessionKeys}</td><td>comma separated list of keys of session variables that should be returned to caller, for correct results as well as for erronous results. (Only for listeners that support it, like JavaListener)</td><td>&nbsp;</td></tr>
+ * <tr><td>{@link #setMultipartResponse(boolean) multipartResponse}</td><td>currently used to mimic the HttpSender when it is stubbed locally. It could be useful in other situations too although currently the response string is used which isn't streamed, it would be better to pass the multipart as an input stream in the context map in which case content type and charset could also be passed</td><td>false</td></tr>
+ * <tr><td>{@link #setMultipartResponseContentType(String) multipartResponseContentType}</td><td></td><td>application/octet-stream</td></tr>
+ * <tr><td>{@link #setMultipartResponseCharset(String) multipartResponseCharset}</td><td></td><td>UTF-8</td></tr>
  * </table>
  * </p>
  * Any parameters are copied to the PipeLineSession of the service called.
@@ -64,10 +69,12 @@ import org.apache.commons.lang.StringUtils;
  * @since   4.4.5
  */
 public class IbisJavaSender extends SenderWithParametersBase implements HasPhysicalDestination {
-	
 	private String name;
 	private String serviceName;
-	private String returnedSessionKeys=null;
+	private String returnedSessionKeys = null;
+	private boolean multipartResponse = false;
+	private String multipartResponseContentType = "application/octet-stream";
+	private String multipartResponseCharset = "UTF-8";
 
 	public void configure() throws ConfigurationException {
 		super.configure();
@@ -91,10 +98,16 @@ public class IbisJavaSender extends SenderWithParametersBase implements HasPhysi
 			if (paramList!=null) {
 				context = prc.getValueMap(paramList);
 			} else {
-				context=new HashMap();			
+				context=new HashMap();
 			}
 			DispatcherManager dm = DispatcherManagerFactory.getDispatcherManager();
 			result = dm.processRequest(getServiceName(),correlationID, message, context);
+			if (isMultipartResponse()) {
+				return HttpSender.handleMultipartResponse(multipartResponseContentType,
+						new ByteArrayInputStream(result.getBytes(multipartResponseCharset)),
+						prc, null);
+			}
+		
 		} catch (ParameterException e) {
 			throw new SenderException(getLogPrefix()+"exception evaluating parameters",e);
 		} catch (Exception e) {
@@ -110,14 +123,12 @@ public class IbisJavaSender extends SenderWithParametersBase implements HasPhysi
 		return result;
 	}
 
-
 	public void setName(String name) {
 		this.name=name;
 	}
 	public String getName() {
 		return name;
 	}
-
 
 	public String getServiceName() {
 		return serviceName;
@@ -130,8 +141,17 @@ public class IbisJavaSender extends SenderWithParametersBase implements HasPhysi
 	public void setReturnedSessionKeys(String string) {
 		returnedSessionKeys = string;
 	}
+
 	public String getReturnedSessionKeys() {
 		return returnedSessionKeys;
+	}
+
+	public void setMultipartResponse(boolean b) {
+		multipartResponse = b;
+	}
+
+	public boolean isMultipartResponse() {
+		return multipartResponse;
 	}
 
 }