+++ RELEASES.md
@@ -110,6 +110,7 @@ Upcoming
 - Add facility to generate flow images by CommandSender next to HttpSender
 - Bugfix "facility to hide properties" caused by splitting configuration
 - Introduction of CounterSwitchPipe (for load balancing between pipes)
+- Add mime multipart response support to http sender